Split git/index utils into separate fiels
