+++ bin-src/tasks/gitInfo.ts
@@ -1,13 +1,9 @@
 import picomatch from 'picomatch';
 
 import getCommitAndBranch from '../git/getCommitAndBranch';
-import {
-  getParentCommits,
-  getBaselineBuilds,
-  getChangedFiles,
-  getSlug,
-  getVersion,
-} from '../git/git';
+import { getChangedFiles, getSlug, getVersion } from '../git/git';
+import { getParentCommits } from '../git/getParentCommits';
+import { getBaselineBuilds } from '../git/getBaselineBuilds';
 import { exitCodes, setExitCode } from '../lib/setExitCode';
 import { createTask, transitionTo } from '../lib/tasks';
 import { matchesFile } from '../lib/utils';
